Revert "Liaison mercure front"
