+++ .idea/testAlpine.iml
@@ -154,10 +154,6 @@
 <module type="WEB_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/spec" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/src/src" isTestSource="false" packagePrefix="App\" />
-      <sourceFolder url="file://$MODULE_DIR$/src/tests" isTestSource="true" packagePrefix="App\Tests\" />
-      <sourceFolder url="file://$MODULE_DIR$/tests" isTestSource="true" />
       <excludeFolder url="file://$MODULE_DIR$/fnacApiConnector/vendor/composer" />
       <excludeFolder url="file://$MODULE_DIR$/fnacApiConnector/vendor/symfony/yaml" />
       <excludeFolder url="file://$MODULE_DIR$/fnacApiConnector/vendor/symfony/polyfill-ctype" />
