+++ front/src/Component/Chat/Conversation.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
 import MessageInput from './MessageInput'
-import ChatMercure from './ChatMercure'
+import SideCard from './SideCard'
+import UserSearchbar from './UserSearchbar'
 import UpperBar from './UpperBar'
 
 const StyledConversation = styled.div`
@@ -13,7 +14,6 @@ const Conversation = () => {
     return (
         <StyledConversation>
             <UpperBar/>
-            <ChatMercure/>
             <MessageInput/>
         </StyledConversation>
     )
