+++ front/src/Container/Chat.jsx
@@ -1,5 +1,6 @@
 import {useState} from "react";
 import '../assets/css/chat.css';
+import {useLocation, useNavigate} from "react-router-dom";
 import NavBar from "../Component/Global/NavBar";
 import ConvSideBar from "../Component/Chat/ConvSideBar";
 import Conversation from "../Component/Chat/Conversation";
