+++ front/src/main.jsx
@@ -5,6 +5,8 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import './assets/css/global.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
     <App />
+  </React.StrictMode>
 )
