+++ src/.env
@@ -39,7 +39,7 @@ MERCURE_JWT_SECRET=mkjgfdsmlgfjfjdglkfdlkjgflkdjgjdslgkfdlksjgfdsglkfdjslkj
 ###< symfony/mercure-bundle ###
 
 ###> nelmio/cors-bundle ###
-CORS_ALLOW_ORIGIN='^http?://(localhost|127\.0\.0\.1)(:[0-9]+)?$'
+CORS_ALLOW_ORIGIN='^https?://(localhost|127\.0\.0\.1)(:[0-9]+)?$'
 ###< nelmio/cors-bundle ###
 ###> lexik/jwt-authentication-bundle ###
 JWT_SECRET_KEY=%kernel.project_dir%/config/jwt/private.pem
