+++ src/public/index.php
@@ -2,11 +2,6 @@
 
 use App\Kernel;
 
-header('Access-Control-Allow-Origin:http://localhost:5173');
-header('Access-Control-Allow-Headers:*');
-header('Access-Control-Allow-Credentials:true');
-header('Access-Control-Allow-Headers:X-Requested-With, Content-Type, withCredentials');
-
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
 return function (array $context) {
